requirements.txt
black==21.12b0
PyGithub==1.55      # utilities for parsing github
pytest==6.2.5
pytest-xdist==2.5.0
requests==2.26.0
